Adjust clr loading and merge Mono code into one file
