+++ pythonnet/.gitignore
@@ -1,0 +1,3 @@
+mono/
+netfx/
+runtime/
