+++ pythonnet/__init__.py
@@ -1,0 +1,3 @@
+def get_assembly_path():
+    import os
+    return os.path.dirname(__file__) + "/runtime/Python.Runtime.dll"
